Add: CSV files to generate a Key Resources Table (STAR Methods)
Why: a framework for Structured, Transparent, Accessible Reporting
(STAR) methods was developed by Cell. Part of the framework
involves supplying a list of all items used in the course of the project
and a list of where raw data has been deposited.

How: Created CSV files following the template provided by cell
https://www.cell.com/star-authors-guide
There may be a goal to create an automated table based on the CSV files
in the future.

Tags: STAR methods, protocol, Cell