Add: Quatro site under ./site
Why: Want to use an easier format (markdown) while also having the
ability to do computations and display latex

How: For now, this is going to be using a local render and then push.
In the future, it will be a push followed by a CI GitHub Action.
The action is already configured, but I am considering how best to
run this.

Tags: quatro, latex, html, site, github pages